% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generate-truth.R
\name{shapes_to_muphi}
\alias{shapes_to_muphi}
\alias{muphi_to_shapes}
\title{Convert between parametrizations of the Beta distribution}
\usage{
shapes_to_muphi(shape1, shape2)

muphi_to_shapes(mu, phi)
}
\arguments{
\item{shape1}{The first shape parameter of the Beta distribution.}

\item{shape2}{The second shape parameter of the Beta distribution.}

\item{mu}{The mean of the Beta distribution.}

\item{phi}{The precision of the Beta distribution.}
}
\value{
A list with two elements containing either the converted shape
parameters or mean and precision values.
}
\description{
Given two shape parameters, calculate the mean and precision of the Beta
distribution; or, given a mean and precision, calculate the corresponding
shape parameters.
}
\details{
The Beta distribution is defined by two shape parameters, α and
β. Often it is desirable to parametrize the Beta distribution in terms
of its mean and precision, rather than the shape parameters. Given, shape
parameters, the mean (μ) and precision (φ) are calculated as:

\deqn{\mu = \frac{\alpha}{\alpha + \beta}}
\deqn{\phi = \alpha + \beta}

Similarly, given a mean and precision for the Beta distribution, the shape
parameters can be calculated as:

\deqn{\alpha = \mu\phi}
\deqn{\beta = (1 - \mu)\phi}

For an intuitive introduction to the Beta distribution, see this
\href{https://www.andrewheiss.com/blog/2021/11/08/beta-regression-guide/}{post}
from Andrew Heiss.
}
\examples{
shapes_to_muphi(6, 4)

muphi_to_shapes(0.6, 10)

}
\author{
Andrew Heiss
}
\concept{Beta}
