% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal-methods.R
\name{calc_gmean}
\alias{calc_gmean}
\title{G-mean method}
\usage{
calc_gmean(estimates, truth)
}
\arguments{
\item{estimates}{A vector of classification probabilities. Values should
represent the probability of \code{1} in the \code{truth} argument.}

\item{truth}{An integer vector of \code{0} and \code{1} representing the true
classifications.}
}
\value{
A numeric scalar representing the optimal probability threshold.
}
\description{
Calculate the optimal probability classification threshold using the
G-Mean method.
}
\details{
The G-mean method (Kubat & Matwin, 1997) is defined as the square root of the
product of sensitivity and specificity at a given threshold. The optimal
threshold is the threshold with the greatest g-mean.

The optimality criterion is then defined as:

\deqn{\text{max}\sqrt{(sensitivity * specificity)}}
}
\examples{
calc_gmean(estimates = dcm_probs$att1$estimate,
            truth = dcm_probs$att1$truth)

calc_gmean(estimates = dcm_probs$att2$estimate,
            truth = dcm_probs$att2$truth)

calc_gmean(estimates = dcm_probs$att3$estimate,
            truth = dcm_probs$att3$truth)

}
\references{
Kubat, M. & Matwin, S. (1997, July 8-12). \emph{Addressing the curse
of imbalanced training sets: One-sided selection} [Paper presentation].
International Conference on Machine Learning, Nashville, TN.
}
\seealso{
Other optimal threshold methods:
\code{\link[=calc_cz]{calc_cz()}},
\code{\link[=calc_topleft]{calc_topleft()}},
\code{\link[=calc_youden]{calc_youden()}}
}
\concept{optimal threshold methods}
