% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_volume.R
\name{calculate_volume}
\alias{calculate_volume}
\title{Calculate volume of an ellipsoid}
\usage{
calculate_volume(width, length)
}
\arguments{
\item{width}{numeric; width of object}

\item{length}{numeric; length of object}
}
\value{
The rounded object volume
}
\description{
This functions calculates the volume of an object given width and length
assuming an ellipsoid shape with equal width and height.
}
