% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateBICvalues.R
\name{calculateBICvalues}
\alias{calculateBICvalues}
\title{Calculate the BIC values}
\usage{
calculateBICvalues(data, k = 3, parallel = FALSE)
}
\arguments{
\item{data}{dataframe of normalized ADT values: Cells as rows and Protein as columns}

\item{k}{maximum number of underlying components (cell populations) to consider fitting through Gaussian Mixture Models (3 by default)}

\item{parallel}{to run the process in parallel (FALSE by default)}
}
\value{
Returns a list, `bic_values` of vectors each corresponding to a column in the input dataframe:

* each vector has k number observations corresponding to k=1 to k=k (maximum number of underlying components)
}
\description{
This function calculates BIC values for each fitting from k=1 to k=k using Mclust function from
mclust R package. The higher BIC value points to better fit through mclust package.
}
\examples{
data("adt.df")
calculateBICvalues(adt.df, k=3, parallel = FALSE)

}
