% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{create_roc}
\alias{create_roc}
\title{Calculate a receiver operating characteristic curve}
\usage{
create_roc(estimates, truth)
}
\arguments{
\item{estimates}{A vector of classification probabilities. Values should
represent the probability of \code{1} in the \code{truth} argument.}

\item{truth}{An integer vector of \code{0} and \code{1} representing the true
classifications.}
}
\value{
An \link[yardstick:roc_curve]{roc_df} object.
}
\description{
Construct the full ROC curve given probability estimates and true
classifications.
}
\examples{
create_roc(estimates = dcm_probs$att1$estimate, truth = dcm_probs$att1$truth)
}
