% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllPossibleFits.R
\name{getAllPossibleFits}
\alias{getAllPossibleFits}
\title{Get All Possible Fits}
\usage{
getAllPossibleFits(bic_values, margin_thr)
}
\arguments{
\item{bic_values}{a list of vectors, each contatining BIC values for k number of fittings for each protein distribution}

\item{margin_thr}{a minimum margin threshold for BIC of k=i+1 to be greater than BIC of k=i (BIC(k=i+1)>(1+margin_thr)*BIC(k=i))}
}
\value{
Returns a list, `k_list` of vectors:

* each vector contains values for possible components fitted best between k=1 to k=k
}
\description{
This function extracts the possible fits for distribution of each protein, by evaluating in each pair of subsequent
fittings whether k=i+1 has greater BIC value than k=i.
}
\examples{
data("adt.df")
GetAllPossibleFits(bic_values, margin_thr=0.1)


}
