% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_fittings_thr.R
\name{publish_fittings_thr}
\alias{publish_fittings_thr}
\title{Save the fitted plots for each possible k in surface marker distributions}
\usage{
publish_fittings_thr(fittings, k_list, thresholds, path = path, seed = 42)
}
\arguments{
\item{fittings}{A list of fitted parameters for each distribution (component) under each k across all markers.}

\item{k_list}{A list with set of possible k values for each marker as an element.}

\item{thresholds}{A data frame with calculated thresholds.}

\item{path}{A path to the directory in which the plots will be saved.}
}
\value{
A data frame with calculated thresholds: bi_thr1 defines the mean+3SD of the first component and bi_cut refers to the cutpoint on x axis where the two components meet in a bimodal (k=2) distribution. tri_thr1 and tri_thr2 indicate mean+3SD of the first and second components in a trimodal (k=3) distribution. Subsequently, tri_cut1 and tri_cut2 represent the cutpoints between first and second component, and second and third components, repectively, in a trimodal distribution. The user should choose which thereshold to use the recomended value is mean+3SD of the first component in k=2.
}
\description{
Save the fitted plots for each possible k in surface marker distributions
}
\examples{
publish_fittings_thr(fittings = fittings, k_list = k.list, thresholds = thresholds, path = '/Users/oliaei/Projects/Packages/ThresholdR/Vignette/02Plots/', seed = 42)
}
