% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_survival.R
\name{survival_custom}
\alias{survival_custom}
\title{Custom survival plots
This function is modified after the function in '~/src/functions.R', and is
partly inspired by the Imvigor210Corebiologies R package (Mariathasan et al., Nature, 2018).}
\usage{
survival_custom(survFit, filename = NULL, title = NULL,
  custom_palette = NULL, custom_legends = NULL, width = 2,
  height = 2.2, xlab = "Overall survival (years)",
  ylab = "Survival probability")
}
\arguments{
\item{survFit}{survFit object}

\item{filename}{character or \code{NULL}; define filename when exporting pdf}

\item{title}{character; plot title}

\item{custom_palette}{character; custom color palette}

\item{custom_legends}{character; custom legend labels (take note of warning below)}

\item{width}{numeric; width of pdf export}

\item{height}{numeric; height of pdf export}

\item{xlab}{character; x-axis label}

\item{ylab}{character; y-axis label}
}
\description{
Custom survival plots
This function is modified after the function in '~/src/functions.R', and is
partly inspired by the Imvigor210Corebiologies R package (Mariathasan et al., Nature, 2018).
}
\author{
Benjamin Ostendorf
}
