% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdr_functions.R
\name{time_to_one_threshold}
\alias{time_to_one_threshold}
\title{Calculate latency and status for meeting threshold}
\usage{
time_to_one_threshold(dataset, object, identifier = UID, time = age,
  group = group, size_threshold)
}
\arguments{
\item{dataset}{matrix}

\item{object}{(unquoted) variable in 'dataset' to calculate threshold time for}

\item{identifier}{(unquoted) variable in 'dataset' defining samples/replicates;
one sample can have several pertaining objects}

\item{time}{(unquoted) variable in 'dataset' indicating time at measurement}

\item{group}{(unquoted) variable in 'dataset' indicating groups samples pertain to}

\item{size_threshold}{numeric; set threshold size}
}
\value{
matrix containing time and status (threshold reached/censored)
}
\description{
This function takes a matrix with longitudinal measurements (e.g., volumes) and
timepoints as input and returns a matrix with status (given threshold reached vs.
censored) and latency.
}
\author{
Benjamin Ostendorf
}
