% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal-methods.R
\name{calc_cz}
\alias{calc_cz}
\title{Concordance probability method}
\usage{
calc_cz(estimates, truth)
}
\arguments{
\item{estimates}{A vector of classification probabilities. Values should
represent the probability of \code{1} in the \code{truth} argument.}

\item{truth}{An integer vector of \code{0} and \code{1} representing the true
classifications.}
}
\value{
A numeric scalar representing the optimal probability threshold.
}
\description{
Calculate the optimal probability classification threshold using the
concordance probability method (CZ).
}
\details{
The concordance probability method (Liu, 2012) is defined as the product of
the sensitivity and specificity at a given threshold cut-point. By utilizing
the product of both metrics, the value for the cut-point remains within the
range of [0, 1]. The optimal threshold is the cut-point that maximizes
the area of the rectangle related to the ROC curve.

The optimality criterion is then defined as:

\deqn{\text{max}(sensitivity * specificity)}
}
\examples{
calc_cz(estimates = dcm_probs$att1$estimate,
            truth = dcm_probs$att1$truth)

calc_cz(estimates = dcm_probs$att2$estimate,
            truth = dcm_probs$att2$truth)

calc_cz(estimates = dcm_probs$att3$estimate,
            truth = dcm_probs$att3$truth)

}
\references{
Liu, X. (2012). Classification accuracy and cut point selection.
\emph{Statistics in Medicine, 31}(23), 2676-2686.
\doi{10.1002/sim.4509}
}
\seealso{
Other optimal threshold methods: 
\code{\link{calc_gmean}()},
\code{\link{calc_topleft}()},
\code{\link{calc_youden}()}
}
\concept{optimal threshold methods}
