% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal-methods.R
\name{calc_youden}
\alias{calc_youden}
\title{Youden's J statistic}
\usage{
calc_youden(estimates, truth)
}
\arguments{
\item{estimates}{A vector of classification probabilities. Values should
represent the probability of \code{1} in the \code{truth} argument.}

\item{truth}{An integer vector of \code{0} and \code{1} representing the true
classifications.}
}
\value{
A numeric scalar representing the optimal probability threshold.
}
\description{
Calculate the optimal probability classification threshold using Youden's J
statistic.
}
\details{
Youden's J statistic (Youden, 1950) is defined as \code{sensitivty} +
\code{specificity} - 1. As such, the value of the J-index ranges from [0, 1],
and is 1 when there are no false positive and no false negatives (i.e.,
sensitivity and specificity). The J-index is calculated for each possible
threshold defined by the ROC curve (e.g., \code{\link[=create_roc]{create_roc()}}).  The optimality
criterion is then defined as:

\deqn{\text{max}(sensitivity + specificity - 1)}
}
\examples{
calc_youden(estimates = dcm_probs$att1$estimate,
            truth = dcm_probs$att1$truth)

calc_youden(estimates = dcm_probs$att2$estimate,
            truth = dcm_probs$att2$truth)

calc_youden(estimates = dcm_probs$att3$estimate,
            truth = dcm_probs$att3$truth)

}
\references{
Youden, W. J. (1950). Index for rating diagnostic tests.
\emph{Cancer, 3}(1), 32-35.
\doi{10.1002/1097-0142(1950)3:1<32::AID-CNCR2820030106>3.0.CO;2-3}
}
\seealso{
Other optimal threshold methods:
\code{\link[=calc_cz]{calc_cz()}},
\code{\link[=calc_gmean]{calc_gmean()}},
\code{\link[=calc_topleft]{calc_topleft()}}
}
\concept{optimal threshold methods}
