% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdr_functions.R
\name{calculate_volume}
\alias{calculate_volume}
\title{Calculate object volume}
\usage{
calculate_volume(width, length)
}
\arguments{
\item{width}{numeric; width of object}

\item{length}{numeric; length of object}
}
\value{
The rounded object volume
}
\description{
This functions calculates the volume of an object given width and length
using the formula for an ellipsoid.
}
\author{
Benjamin Ostendorf
}
