% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_k.R
\name{fit_k}
\alias{fit_k}
\title{Fit for the possible number of unrlying components (k) across all markers}
\usage{
fit_k(
  data,
  k_list,
  margin.den = 0.1,
  epsilon = 0.01,
  maxit = 10000,
  maxrestarts = 20,
  seed = 42
)
}
\arguments{
\item{data}{A data frame of CLR normalized ADT expression values. Cells are in the rows and ADT probes (surface markers) are in columns.}

\item{k_list}{A list of the possible number of underlying components across all markers.}

\item{margin.den}{The margin for difference between the density of the mean in the first fitted component and the density value of the input data at x=mean(C1). This is needed to correct the underestimated standard deviation of the first component which happens sometimes with normalmixEM function from mixtools package.}

\item{maxit}{Maximum number of iterations for fitting of each k. The argument come from the 'mixtools' package.}

\item{maxrestarts}{Maximum number of restarts for fitting of each k. The argument come from the 'mixtools' package.}

\item{seed}{Random number generator. Is set to a user provided value to promote reproducibility.}
}
\value{
A list of parameters such as mean, standard-deviation, weight (lambda) for each component in different k values across the markers. This will later be used to define thresholds.
}
\description{
Fit for the possible number of unrlying components (k) across all markers
}
\examples{
fittings <- fit_k(data = adt, k_list = k.list, margin.den = 0.1, epsilon = 0.01, maxit = 1000, maxrestarts = 10)
}
