% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-truth.R
\name{generate_truth}
\alias{generate_truth}
\title{Generate binary values from a vector of probabilities}
\usage{
generate_truth(estimates)
}
\arguments{
\item{estimates}{A vector of probabilities}
}
\value{
An integer vector of \code{0} and \code{1}, the same length as \code{estimate}.
}
\description{
For each probability a binary outcome is created using \code{\link[=rbern]{rbern()}}. In
practice, it is likely useful generate multiple truth values that can be
tested.
}
\examples{
generate_truth(runif(10))

generate_truth(runif(5))
}
