% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thresholds.R
\name{get_thresholds}
\alias{get_thresholds}
\title{Get Thresholds}
\usage{
get_thresholds(fittings, k_list)
}
\arguments{
\item{fittings}{dataframe of normalized ADT values: Cells as rows and Protein as columns}

\item{k_list}{a list of vectors, each vector contains values for possible components fitted best between k=1 to k=k for normalized distribution of each protein}
}
\value{
a dataframe, `thresholds` including columns for thresholds of each fitting, namely:

* `bimodal_thr` - thresold for bimodal (k=2) fittings
* `trimodal_thr` - thresold for trimodal (k=3) fittings
}
\description{
This function calculates the threshold based on mean and standard deviation of the
first component (M1+3SD)
}
\examples{
data("adt.df")
thresholds.test <- get_thresholds(fittings = fittings.test, k_list = k_list.test)
#' @export

}
