% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latency_two_objects.R
\name{latency_two_thresholds}
\alias{latency_two_thresholds}
\title{Calculate latency and status between thresholds}
\usage{
latency_two_thresholds(dataset, object, identifier = UID, time = age,
  group = group, lower_threshold, upper_threshold)
}
\arguments{
\item{dataset}{matrix}

\item{object}{(unquoted) variable in 'dataset' to calculate latency for}

\item{identifier}{(unquoted) variable in 'dataset' defining samples/replicates;
one sample can have several pertaining objects}

\item{time}{(unquoted) variable in 'dataset' indicating time at measurement}

\item{group}{(unquoted) variable in 'dataset' indicating groups samples pertain to}

\item{lower_threshold}{numeric; lower threshold}

\item{upper_threshold}{numeric; upper threshold}
}
\description{
This function takes as input a matrix with longitudinal measurements (e.g., volume)
and timepoints and returns a matrix holding the status (second threshold
reached vs. censored) and latency between two given thresholds.
}
