% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal-iterate.R
\name{optimal_iterate}
\alias{optimal_iterate}
\title{Use iteration to estimate an optimal probability threshold}
\usage{
optimal_iterate(
  estimates,
  weight_method,
  optimal_method,
  ...,
  additional_criterion = NULL,
  iter_burnin = 100,
  iter_retain = 1000,
  comp_thresholds = NULL,
  metrics = NULL
)
}
\arguments{
\item{estimates}{A vector of probabilities.}

\item{weight_method}{The method for generating classifications, weighted by
the current estimate of the optimal threshold. One of
"beta", "distance".}

\item{optimal_method}{The method for estimating the optimal threshold. One of
"youden", "topleft", "cz", "gmean".}

\item{...}{Additional arguments passed to the corresponding weighting method.}

\item{additional_criterion}{Optional. If provided, must be a class
probability metric from \link[yardstick:yardstick-package]{yardstick}.}

\item{iter_burnin}{The number of iterations to run and then discard (see
Details below).}

\item{iter_retain}{The number of iterations to retain (see Details below).}

\item{comp_thresholds}{Additional threshold values to evaluate against the
average optimal threshold (e.g., to compare the optimal threshold to a
competing threshold such as 0.5). If \code{NULL} (the default), no additional
thresholds are included in the performance evaluation.}

\item{metrics}{Either \code{NULL} or a \code{\link[yardstick:metric_set]{yardstick::metric_set()}} with a list of
performance metrics to calculate. The metrics should all be oriented
towards hard class predictions (e.g., \code{\link[yardstick:sens]{yardstick::sensitivity()}},
\code{\link[yardstick:accuracy]{yardstick::accuracy()}}, \code{\link[yardstick:recall]{yardstick::recall()}}) and not class
probabilities. A set of default metrics is used when \code{NULL} (see
\code{\link[probably:threshold_perf]{probably::threshold_perf()}} for details).}
}
\value{
A \link[tibble:tibble-package]{tibble} with 1 row per threshold. The
columns are:
\itemize{
\item \code{.threshold}: The optimal threshold.
\item If \code{additional_criterion} was specified, an \link[posterior:rvar]{rvar}
containing the distribution of class probability metrics across all
retained iterations.
\item A set of \link[posterior:rvar]{rvar} objects for each of the specified
performance metrics, containing the distributions across all retained
iterations (i.e., 1 column per specified metric).
}
}
\description{
Use iteration to estimate an optimal probability threshold when true
classifications are unknown.
}
\details{
Additional details...
}
\examples{
est <- runif(100)
optimal_iterate(estimates = est, weight_method = "distance",
                optimal_method = "youden", iter_retain = 100)
}
\seealso{
Other threshold approximation methods: 
\code{\link{optimal_resample}()}
}
\concept{threshold approximation methods}
