% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_fittings_thr.R
\name{publish_fittings_thr}
\alias{publish_fittings_thr}
\title{Publish Fittings with Assigned Thresholds}
\usage{
publish_fittings_thr(fittings, k_list, thresholds, path = path)
}
\arguments{
\item{fittings}{a list of fittings for each protein, each containg fitting for all possible fits}

\item{k_list}{a list of vectors, each vector contains values for possible components fitted best between k=1 to k=k for normalized distribution of each protein}

\item{thresholds}{a dataframe of thresholds which has proteins or surface markers as rows and bimodal and/or trimodal thresholds as columns.}

\item{path}{the path you would like to save the fitting density plots.}
}
\value{
It prints the the plots as pdf files in the path provided for path argument.
}
\description{
This function publishes bimodal and trimodal fittings with assigned thresholds in the location of interest.
}
\examples{
data("adt.df")
publish_fittings_thr(fittings = fittings.test, k_list =k_list.test ,thresholds = thresholds.test, path = "~/test_run/figures/")

}
