% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdr-package.R
\docType{package}
\name{thresholdr-package}
\alias{thresholdr}
\alias{thresholdr-package}
\title{thresholdr: What the Package Does (One Line, Title Case)}
\description{
Diagnostic classification models are confirmatory latent class models that provide probabilities of proficiency for respondents on each of the assessed skills (Bradshaw, 2016 \doi{10.1002/9781118956588.ch13}). Often, results based on these models are reported as classifications, using 0.5 as the probability threshold for classification (Bradshaw & Levy, 2019 \doi{10.1111/emip.12247}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thresholdr.measr.info}
  \item \url{https://github.com/r-dcm/thresholdr}
  \item Report bugs at \url{https://github.com/r-dcm/thresholdr/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

Authors:
\itemize{
  \item Jonathan A. Pedroza
}

Other contributors:
\itemize{
  \item University of Kansas [copyright holder]
}

}
\keyword{internal}
