% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdr_functions.R
\name{time_to_one_threshold}
\alias{time_to_one_threshold}
\title{Calculate time at which volume exceeds threshold}
\usage{
time_to_one_threshold(dataset, object, identifier = UID, time = age,
  group = group, size_threshold)
}
\arguments{
\item{dataset}{matrix containing variables "UID", "group", and "age"}

\item{object}{character; variable in 'dataset' to calculate threshold time for}

\item{identifier}{character; variable in 'dataset' defining samples/replicates;
one sample can have several pertaining objects}

\item{time}{character; variable in 'dataset' indicating time at measurement}

\item{group}{character; variable in 'dataset' indicating groups samples pertain to}

\item{size_threshold}{numeric; set threshold size}
}
\value{
matrix containing time and status (threshold reached/censored)
}
\description{
This function calculates the age at which an object volume exceeds a given threshold;
if the endpoint is reached it returns the day and annotates status as "1"; if the endpoint
is not reached the last day of measurement is recorded and status annotated with "0".
}
\author{
Benjamin Ostendorf
}
